+++ crates/nostr/src/nips/mod.rs
@@ -22,6 +22,7 @@ pub mod nip17;
 pub mod nip19;
 pub mod nip21;
 pub mod nip22;
+pub mod nip25;
 pub mod nip34;
 pub mod nip35;
 pub mod nip38;